Add special case for chance of dam >= 0
This was displaying as chance for strictly over 0 damage, i.e., chance
to do any positive damage, since we weren't counting fail rolls at all
if they fell underneath the threshold to contribute non-zero damage.

It's maybe a little hacky to just special case this at the start of
finding the chance. But it works and saves some pointless computation.