Delete init.txt
It's a pain, error-prone, and unnecessary to try to keep the output file
in lock-step with the parts, and it's confusing to keep in the repo a
version of the file that can always be out-of-date with the files that
construct it. It's simple to construct init.txt, so just do that when
you want it.